src/requirements.txt
flask
python-dotenv
setuptools
requests
PyJWT  # TOKEN

# DB
Flask-Migrate
flask-sqlalchemy
flask_validator 

psycopg2-binary

# helpers
pydash
# pprint

# For Serializing Objects
marshmallow  
marshmallow_sqlalchemy
